Conformed to Clean Architecture
