+++ src/main/java/recipe/foodbar/config/ManualConfig.java
@@ -7,8 +7,8 @@ import recipe.foodbar.repository.simpleDB.InMemoryRecipeRepository;
 import recipe.foodbar.repository.simpleDB.InMemoryUserRepository;
 import recipe.foodbar.usecase.cuisine.CreateCuisine;
 import recipe.foodbar.usecase.cuisine.port.CuisineRepository;
-import recipe.foodbar.usecase.recipe.CreateRecipe;
-import recipe.foodbar.usecase.recipe.UpdateRecipe;
+import recipe.foodbar.usecase.recipe.manager.CreateRecipe;
+import recipe.foodbar.usecase.recipe.manager.UpdateRecipe;
 import recipe.foodbar.usecase.recipe.port.RecipeRepository;
 import recipe.foodbar.usecase.review.CreateReview;
 import recipe.foodbar.usecase.user_example.CreateUser;
