+++ src/main/java/recipe/foodbar/controller/recipe/EditRecipe.java
@@ -1,14 +1,14 @@
 package recipe.foodbar.controller.recipe;
 
-import recipe.foodbar.usecase.recipe.RecipeRequestModel;
-import recipe.foodbar.usecase.recipe.edit.RecipeEditor;
+import recipe.foodbar.usecase.recipe.ds.RecipeRequestModel;
+import recipe.foodbar.usecase.recipe.manager.edit.IRecipeEditor;
 
 /**
  * A controller to edit a recipe.
  */
 public class EditRecipe {
     private final RecipeRequestModel rrm;
-    private final RecipeEditor e;
+    private final IRecipeEditor e;
 
     /**
      * Constructor for EditRecipe
@@ -16,7 +16,7 @@ public class EditRecipe {
      * @param rrm request to edit.
      * @param e   An editor interface.
      */
-    public EditRecipe(RecipeRequestModel rrm, RecipeEditor e) {
+    public EditRecipe(RecipeRequestModel rrm, IRecipeEditor e) {
         this.rrm = rrm;
         this.e = e;
