+++ src/main/java/recipe/foodbar/presenter/RecipePresenter.java
@@ -1,6 +1,6 @@
-package recipe.foodbar.usecase.recipe.edit;
+package recipe.foodbar.presenter;
 
-import recipe.foodbar.usecase.recipe.RecipeResponseModel;
+import recipe.foodbar.usecase.recipe.ds.RecipeResponseModel;
 import recipe.foodbar.usecase.recipe.port.RecipeOut;
 
 public class RecipePresenter implements RecipeOut {
