--- src/main/java/recipe/foodbar/usecase/recipe/edit/Editor.java
@@ -1,4 +1,0 @@
-package recipe.foodbar.usecase.recipe.edit;
-
-public interface Editor {
-}
