+++ src/main/java/recipe/foodbar/usecase/recipe/manager/edit/Editor.java
@@ -1,0 +1,4 @@
+package recipe.foodbar.usecase.recipe.manager.edit;
+
+public interface Editor {
+}
