+++ src/main/java/recipe/foodbar/usecase/recipe/manager/sort/CuisineComparator.java
@@ -1,17 +1,14 @@
-package recipe.foodbar.usecase.recipe.sort;
+package recipe.foodbar.usecase.recipe.manager.sort;
 
 import recipe.foodbar.entities.Recipe;
 
-/**
- * This comparator class is responsible for comparing two recipes and returning an ordering based on rating
- * (used to reorder by rating)
- */
-public class RatingComparator extends AbstractComparator {
+public class CuisineComparator extends AbstractComparator {
 
     /**
      * Compares its two arguments for order.  Returns a negative integer,
      * zero, or a positive integer as the first argument is less than, equal
      * to, or greater than the second.
+     *
      * @param r1 the first object to be compared.
      * @param r2 the second object to be compared.
      * @return a negative integer, zero, or a positive integer as the
@@ -24,9 +21,6 @@ public class RatingComparator extends AbstractComparator {
      */
     @Override
     public int compare(Recipe r1, Recipe r2) {
-        /*
-        Comparable recipe?
-         */
-        return r1.getRating() - r2.getRating();
+        return String.CASE_INSENSITIVE_ORDER.compare(r1.getCuisineId(), r2.getCuisineId());
     }
 }
