+++ src/main/java/recipe/foodbar/usecase/recipe/manager/sort/Sorter.java
@@ -1,0 +1,5 @@
+package recipe.foodbar.usecase.recipe.manager.sort;
+
+public interface Sorter {
+
+}
