+++ src/main/java/recipe/foodbar/usecase/recipe/port/RecipeOut.java
@@ -1,7 +1,10 @@
 package recipe.foodbar.usecase.recipe.port;
 
-import recipe.foodbar.usecase.recipe.RecipeResponseModel;
+import recipe.foodbar.usecase.recipe.ds.RecipeResponseModel;
 
+/**
+ * The output boundary for recipes.
+ */
 public interface RecipeOut {
     public default void displayEdited(RecipeResponseModel.ResponseDataType type) {}
 }
