--- src/main/java/recipe/foodbar/usecase/recipe/sort/Sorter.java
@@ -1,5 +1,0 @@
-package recipe.foodbar.usecase.recipe.sort;
-
-public interface Sorter {
-
-}
